# ticketing_service/swagger.yaml
openapi: 3.0.3
info:
  version: "0.1"
  title: "TickTap Ticketing Service API"
  description: "This is the API documentation for a fictional company called TickTap LLC that services event ticketing for personal and commercial events."
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/zfanelle/ticktap/0.1
  - description: SwaggerHub API Auto Mocking
    url: localhost
paths:
  /ticket:
    get:
      tags:
        - "ticket"
      summary: Get all tickets
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ticket"
    post:
      tags:
        - "ticket"
      summary: Create ticket(s)
      responses:
        "200":
          description: success
      requestBody:
        $ref: "#/components/requestBodies/ticket"
  /ticket/{ticketId}:
    get:
      tags:
        - "ticket"
      summary: Get a ticket by ticket id
      parameters:
        - $ref: "#/components/parameters/ticketId"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ticket"
  /ticket/transaction/{transactionId}:
    get:
      tags:
        - "ticket"
      summary: Get all tickets by transaction id
      parameters:
        - $ref: "#/components/parameters/transactionId"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ticket"

components:
  schemas:
    ticket:
      type: object
      properties:
        ticketId:
          type: integer
        accountId:
          type: integer
        event:
          type: integer
        dateOfPurchase:
          type: string
          format: datetime
    createTicket:
      type: object
      properties:
        transactionId:
          type: integer

  requestBodies:
    ticket:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/createTicket"

  parameters:
    ticketId:
      in: path
      name: ticketId
      required: true
      schema:
        type: integer
    transactionId:
      in: path
      name: transactionId
      required: true
      schema:
        type: integer
tags:
  - name: "transaction"
    description: "Transaction refers to one or more tickets. For example, a family bys 5 tickets to a basketball game in a transaction."